requirements.txt
# Core NLP для русского языка
razdel>=0.3.0
pymorphy2>=0.1.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configs
pyyaml>=6.0

# Visualization
wordcloud>=1.9.2
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
